split aux model functions to a separate module, write metrics to json file, refactor configs
