+++ ml_project/configs/estimator/random-forest.yaml
@@ -1,4 +1,5 @@
 model_artifact_path: ${dir_prefix}/artifacts/random-forest.pkl
+metrics_path: ${dir_prefix}/metrics/random-forest.json
 
 model_type: RandomForest
 pos_label: M
